autonomous_perception_dashboard/requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
